[render] Move files around to make me less insane
